+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "/Users/irahorecka/.virtualenvs/youtube/bin/python"
+}
